python_service/requirements.txt
Flask
sentence-transformers
psycopg2-binary
numpy
requests # For making HTTP requests to the C++ backend